backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
gevent==23.9.1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
requests==2.31.0
firebase-admin==6.4.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-debug-toolbar==4.2.0
